% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{df}
\alias{df}
\title{Calculate the degrees of freedom based on the mean and variance}
\usage{
df(m, var, p)
}
\arguments{
\item{m}{Mean.}

\item{var}{Variance.}

\item{p}{The dimension of the matrix.}
}
\value{
The degrees of freedom of an inverse Wishart distribution.
}
\description{
One can input the expectation and variance of a diagonal element of the matrix.
}
