% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{posterior.Wishart}
\alias{posterior.Wishart}
\title{Posterior based on a Wishart distribution for precision matrix}
\usage{
posterior.Wishart(n, S, w0, U0)
}
\arguments{
\item{n}{Sample size.}

\item{S}{Sample covariance matrix (MLE).}

\item{w0}{Degrees of freedom of the Wishart prior.}

\item{U0}{Scale matrix of the Wishart prior.}
}
\value{
The mean and variance of both Wishart and inverse Wishart distribution.
}
\description{
Combine data and Wishart prior of precision matrix.
}
\seealso{
\code{\link{posterior.iWishart}} covariance matrix.
}
