% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{riWshart}
\alias{riWshart}
\title{Generate random numbers from an inverse Wishart distribution}
\usage{
riWshart(n, df, Sigma)
}
\arguments{
\item{n}{Sample size.}

\item{df}{Degrees of freedom.}

\item{Sigma}{The scale matrix.}
}
\value{
The mean and variance.
}
\description{
This function generates random matrix from an inverse
Wishart distribution.
}
\examples{
n <- 1000
df <- 10
Sigma <- array(c(2,0.5,0.5,1), dim=c(2,2))
dset1 <- rWishart(n, df, Sigma)

apply(dset1, 1:2, mean)  ## mean
apply(dset1, 1:2, var)   ## variance
}
\seealso{
\code{\link{rWishart}} for random matrix from a Wishart distribution.
}
