% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{stat.iWishart}
\alias{stat.iWishart}
\title{Obtain the mean and variance of an inverse Wishart distribution}
\usage{
stat.iWishart(sigma, m)
}
\arguments{
\item{sigma}{The scale matrix.}

\item{m}{Degrees of freedom.}
}
\value{
A 3d array of generated matrices.
}
\description{
This function calculates the mean and variance of an inverse
Wishart distribution based on the input of a scale matrix
and the degrees of freedom.
}
\examples{
sigma <- array(c(2,0.5,0.5,1), dim=c(2,2))
stat.iWishart(sigma, 10)
}
