% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{stat.Wishart}
\alias{stat.Wishart}
\title{Obtain the mean and variance of a Wishart distribution}
\usage{
stat.Wishart(sigma, m)
}
\arguments{
\item{sigma}{The scale matrix.}

\item{m}{Degrees of freedom.}
}
\value{
The mean and variance.
}
\description{
#'This function calculates the mean and variance of a
Wishart distribution based on the input of a scale matrix
and the degrees of freedom.
}
\examples{
sigma <- array(c(2,0.5,0.5,1), dim=c(2,2))
stat.Wishart(sigma, 10)
}
